pyproject.toml
[project]
name = "downing-detect"
version = "1.0.0"
description = "Drowning Detection System - Real-time computer vision system for drowning risk detection"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx==0.27.2",
    "matplotlib>=3.8.0",
    "numpy>=1.21.0,<2.0.0",
    "opencv-python==4.10.0.84",
    "pillow==10.4.0",
    "pydantic==2.9.1",
    "pygame==2.6.1",
    "pyyaml==6.0.1",
    "scipy>=1.12.0",
    "tqdm==4.66.5",
    "ultralytics==8.3.3",
    "lapx>=0.5.2",
    "openai>=2.8.1",
    "rich>=13.9.0",
    "loguru>=0.7.3",
    "prometheus-client==0.19.0",
    "psutil==6.0.0",
    # FastAPI backend dependencies
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "websockets==12.0",
    "python-multipart==0.0.6",
]

[tool.black]
line-length = 100
target-version = ['py312']
